NoticeListViewは通知でしか使わないため削除した
